+++ imports/ui/ThreadItem.vue
@@ -24,10 +24,7 @@ export default {
 .thread-item {
   padding: 12px;
   cursor: default;
-    user-select: none;
-    -moz-user-select: none;
-    -webkit-user-select: none;
-    -ms-user-select: none;
+  user-select: none;
 }
 
 .thread-item:hover {