+++ packages/vue-component/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'akryum:vue-component',
-  version: '0.7.12',
+  version: '0.8.0',
   summary: 'VueJS single-file components that hot-reloads',
   git: 'https://github.com/Akryum/meteor-vue-component',
   documentation: 'README.md'
@@ -30,13 +30,17 @@ Package.registerBuildPlugin({
     'async': '1.4.0',
     'lodash': '4.13.1',
     'vue-template-compiler': '2.1.8',
-    'hash-sum': '1.0.2'
+    'hash-sum': '1.0.2',
+    'source-map': '0.5.6',
+    'source-map-merger': '0.2.0',
+    'generate-source-map': '0.0.5',
+    'autoprefixer': '6.6.0',
   }
 });
 
 Package.onUse(function(api) {
   api.versionsFrom('1.3.3');
   api.use('isobuild:compiler-plugin@1.0.0');
-  api.use('akryum:vue-component-dev-server@0.0.3');
-  api.use('akryum:vue-component-dev-client@0.2.2');
+  api.use('akryum:vue-component-dev-server@0.0.4');
+  api.use('akryum:vue-component-dev-client@0.2.3');
 });